sub-pj/agentic-ai-sales/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sales-presentation-automator"
version = "0.1.0"
description = "AI-powered sales presentation generator CLI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "thedomainai" }
]
dependencies = [
    "click>=8.1.0",
    "anthropic>=0.40.0",
    "python-pptx>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
spa = "spa.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spa"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/spa"
